adding get_data sub

+++ machine_learning/exter_dataset/uitls/download.py
@@ -3,6 +3,7 @@ import tarfile
 import urllib.request
 import zipfile
 from git import Repo
+import git
 from kaggle.api.kaggle_api_extended import KaggleApi
 from datasets import load_dataset
 
@@ -26,15 +27,21 @@ def zip_download(base_path,dataset_name,url,name=None):
     else:
         zip_name = os.path.join(base_path,"repo",dataset_name,dataset_name+".zip")
         dir_name = os.path.join(base_path,"repo",dataset_name)
-    try:
-        print(dir_name)
+    
+    if not os.path.isdir(dir_name):
         os.mkdir(dir_name)
+    try:
+        dir = os.listdir(dir_name)
+        if len(dir) != 0:
+            return
     except:
         pass
-    with urllib.request.urlopen(url) as f:
-        text = f.read()
-        with open(zip_name,"wb") as file :
-            file.write(text)
+    
+    if not os.path.exists(zip_name):
+        with urllib.request.urlopen(url) as f:
+            text = f.read()
+            with open(zip_name,"wb") as file :
+                file.write(text)
     try:
         with zipfile.ZipFile(zip_name,"r") as zip_ref:
             zip_ref.extractall(dir_name)
@@ -51,28 +58,41 @@ def gz_tar_download(base_path,dataset_name,url,name=None):
     else:
         zip_name = os.path.join(base_path,"repo",dataset_name,dataset_name+".gz_tar")
         dir_name = os.path.join(base_path,"repo",dataset_name)
-    try:
+    
+    if not os.path.isdir(dir_name):
         os.mkdir(dir_name)
+    try:
+        dir = os.listdir(dir_name)
+        if len(dir) != 0:
+            return
     except:
         pass
-    with urllib.request.urlopen(url) as f:
-        text = f.read()
-        with open(zip_name,mode="wb") as file :
-            file.write(text)
+    if not os.path.exists(zip_name):
+        with urllib.request.urlopen(url) as f:
+            text = f.read()
+            with open(zip_name,mode="wb") as file :
+                file.write(text)
     with tarfile.open(zip_name) as zip_ref:
         zip_ref.extractall(path=dir_name)
 
 def git_download(base_path,dataset_name,url):
     print("git_download:",dataset_name)
     dir_name = os.path.join(base_path,"repo",dataset_name)
-    try:
+    if not os.path.isdir(dir_name):
         os.mkdir(dir_name)
-    except:
-        pass
-    try:
-        Repo.clone_from(url, dir_name)
-    except:
-        print("fail git_download:",dataset_name)
+    dir = os.listdir(dir_name)
+    if len(dir) == 0:
+        try:
+            Repo.clone_from(url, dir_name)
+        except:
+            print("fail git_download:",dataset_name)
+    else:
+        try:
+            g = git.Git(dir_name)
+            o = g.remotes.origin
+            o.pull()
+        except:
+            print("fail update git:",dataset_name)
 
 
 def kaggle_download(base_path,dataset_name,name):
@@ -86,7 +106,8 @@ def kaggle_download(base_path,dataset_name,name):
 def huggingface_download(base_path,dataset_name,name,subname):
     print("huggingface_download:",dataset_name)
     dir_fs = os.path.join(base_path,"repo", dataset_name)
-    os.mkdir(dir_fs)
+    if not os.path.isdir(dir_fs):
+        os.mkdir(dir_fs)
     try:
         dataset = load_dataset(name, subname,data_dir=dir_fs)
     except:
@@ -97,10 +118,13 @@ def zenodo_download(base_path,dataset_name,id_zenodo):
     print("zenodo_download:",dataset_name)
     from zenodo_client import Zenodo
     dir_fs = os.path.join(base_path,"repo", dataset_name)
-    try:
+    
+    if not os.path.isdir(dir_fs):
         os.mkdir(dir_fs)
-    except:
-        pass
+    
+    dir = os.listdir(dir_fs)
+    if len(dir) != 0:
+        return
     try:
         zenodo = Zenodo()
         dataset = zenodo.download_latest(id_zenodo, path=dir_fs)
@@ -128,14 +152,21 @@ def file_download(base_path,dataset_name,url,name=None,ext=None):
         path_name = os.path.join(base_path,"repo",dataset_name,dataset_name+"."+ext)
     else:
         path_name = os.path.join(base_path,"repo",dataset_name,dataset_name)
+    
     try:
-        os.mkdir(dir_name)
+        dir = os.listdir(dir_name)
+        if len(dir) != 0:
+            return
     except:
         pass
-    with urllib.request.urlopen(url) as f:
-        text = f.read()
-        with open(path_name,mode="wb") as file :
-            file.write(text)
+    
+    if not os.path.isdir(dir_name):
+        os.mkdir(dir_name)
+    if not os.path.exists(path_name):
+        with urllib.request.urlopen(url) as f:
+            text = f.read()
+            with open(path_name,mode="wb") as file :
+                file.write(text)
 
 def drive_google_download(base_path,dataset_name,url):
     pass
@@ -154,13 +185,10 @@ def dropbox_download(base_path,dataset_name,url):
 
 def kaggle_download(base_path,user,dataset_name,file_name):
     print("kaggle_download",dataset_name)
-    try:
-        dir_name = os.path.join(base_path,"repo",dataset_name)
+    dir_name = os.path.join(base_path,"repo",dataset_name)
+    if not os.path.isdir(dir_name):
         os.mkdir(dir_name)
-    except:
-        pass
     dir_name = os.path.join(base_path,"repo",dataset_name,file_name)
-    print(dir_name)
     text= api.datasets_download_file(user,dataset_name,file_name=file_name)
     with open(dir_name,mode="wb") as file :
             file.write(text)
